// aq54-frontend/src/configs/types.ts
export type chartDataType = {
  extT?: number;
  rh?: number;
  o3?: number;
  co?: number;
  no2?: number;
  pm25?: number;
  pm10?: number;
};

export type stationDataType = {
  id?: String;
  station_name?: String;
  timestamp?: String;
  extT?: number;
  rh?: number;
  o3?: number;
  co?: number;
  no2?: number;
  pm25?: number;
  pm10?: number;
};